refactor(meta-api): consolidate schema API implementation
Merge `schema_api_impl.rs` into `schema_api.rs` as trait default
implementations. Convert SchemaApi from separate impl block to
trait with default methods, fixing trait bound redundancy.

- Remove `schema_api_impl.rs` and update lib.rs module exports
- Simplify function signatures using SchemaApi supertrait bounds
- Fix compilation errors in helper function trait constraints
